modify some function prototypes of the dfs_file_ops structure and the function declarations based on it (#7849)

+++ components/drivers/tty/tty.c
@@ -306,9 +306,9 @@ static int tty_ioctl(struct dfs_file *fd, int cmd, void *args)
 }
 
 #ifdef RT_USING_DFS_V2
-static int tty_read(struct dfs_file *fd, void *buf, size_t count, off_t *pos)
+static ssize_t tty_read(struct dfs_file *fd, void *buf, size_t count, off_t *pos)
 #else
-static int tty_read(struct dfs_file *fd, void *buf, size_t count)
+static ssize_t tty_read(struct dfs_file *fd, void *buf, size_t count)
 #endif
 {
     int ret = 0;
@@ -327,9 +327,9 @@ static int tty_read(struct dfs_file *fd, void *buf, size_t count)
 }
 
 #ifdef RT_USING_DFS_V2
-static int tty_write(struct dfs_file *fd, const void *buf, size_t count, off_t *pos)
+static ssize_t tty_write(struct dfs_file *fd, const void *buf, size_t count, off_t *pos)
 #else
-static int tty_write(struct dfs_file *fd, const void *buf, size_t count )
+static ssize_t tty_write(struct dfs_file *fd, const void *buf, size_t count )
 #endif
 {
     int ret = 0;
